renamed Stub -> Fake in gradle, comments & constants everywhere
