+++ app/build.gradle
@@ -48,7 +48,7 @@ android {
 
 dependencies {
     implementation project(':utilities')
-    implementation project(':stubDataGenerator')
+    implementation project(':fakeDataGenerator')
 
     implementation 'androidx.fragment:fragment-ktx:1.6.2'
