+++ settings.gradle
@@ -15,4 +15,4 @@ dependencyResolutionManagement {
 rootProject.name = "HomeTask_ApiPolling"
 include ':app'
 include ':utilities'
-include ':stubDataGenerator'
+include ':fakeDataGenerator'
