+++ stubDataGenerator/src/main/java/com/igor_shaula/fake_data_generator/FakeVehicleGenerator.kt
@@ -1,4 +1,4 @@
-package com.igor_shaula.stub_data_generator
+package com.igor_shaula.fake_data_generator
 
 // TODO this class should produce data similar to what is coming over network - in String/JSON type
